# openapi.yaml
openapi: 3.0.3
info:
  title: SOAR API
  version: 1.0.0
  description: API for SOAR enrichment and response automation

servers:
  - url: / # Typically replaced with your deployed API Gateway URL

paths:
  /:
    get:
      summary: Health check
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /ipabusedb:
    get:
      summary: Lookup IP in AbuseIPDB
      parameters:
        - in: query
          name: ip
          schema:
            type: string
          required: true
          description: IP address to check
      responses:
        '200':
          description: AbuseIPDB lookup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbuseIPDBResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/UpstreamError'

  /anomali:
    get:
      summary: Lookup IOC in Anomali ThreatStream
      parameters:
        - in: query
          name: ioc
          schema:
            type: string
          required: true
          description: Indicator of Compromise (IOC) to check
      responses:
        '200':
          description: Anomali lookup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomaliResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/UpstreamError'

  /datp/upload:
    post:
      summary: Upload IOC to Microsoft Defender (DATP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatpUploadRequest'
      responses:
        '200':
          description: DATP upload result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatpUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /azuread/blockip:
    post:
      summary: Block IP in Azure AD Named Location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AzureAdBlockIpRequest'
      responses:
        '200':
          description: Azure AD block result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AzureAdBlockIpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /sase/block:
    post:
      summary: Block IOC in SASE (Umbrella)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaseBlockRequest'
      responses:
        '200':
          description: SASE block result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaseBlockResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your OAuth 2.0 Bearer token in the format: `Bearer <access_token>`.
        This API requires a valid OAuth 2.0 access token for all endpoints.

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UpstreamError:
      description: Upstream API issue
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    HealthResponse:
      type: object
      properties:
        message:
          type: string
          description: Health check message

    AbuseIPDBResponse:
      type: object
      description: AbuseIPDB lookup result
      properties:
        Confidence:
          type: integer
          description: Abuse confidence score
        Country:
          type: string
          description: Country name
        ReportCount:
          type: integer
          description: Number of reports
        Tor:
          type: boolean
          description: Whether the IP is a Tor exit node
        IOC:
          type: string
          description: The IP address queried
        Private:
          type: boolean
          description: Whether the IP is private

    AnomaliResponse:
      type: object
      description: Anomali lookup result
      properties:
        Confidence:
          type: integer
          description: Average confidence score
        IOCType:
          type: string
          description: Type of Indicator of Compromise
        IOC:
          type: string
          description: The IOC queried
        Grade:
          type: string
          description: Confidence grade (Low, Medium, High)

    DatpUploadRequest:
      type: object
      properties:
        ioc:
          type: string
          description: Indicator of Compromise
        action:
          type: string
          enum: [Block, Audit, BlockAndRemediate]
          description: Action to take
        incident_id:
          type: string
          description: Incident ID
      required: [ioc, action, incident_id]

    DatpUploadResponse:
      type: object
      description: DATP upload result
      properties:
        Added:
          type: boolean
          description: Whether the IOC was added to DATP
        IOC:
          type: string
          description: Indicator of Compromise
        Platform:
          type: string
          description: Platform name (e.g., DATP)
        Action:
          type: string
          description: Action taken (e.g., Block, None)

    AzureAdBlockIpRequest:
      type: object
      properties:
        ioc:
          type: string
          description: IP to block
        list_id:
          type: string
          description: Named location list ID
        list_name:
          type: string
          description: Named location list name
        incident_id:
          type: string
          description: Incident ID
      required: [ioc, list_id, list_name, incident_id]

    AzureAdBlockIpResponse:
      type: object
      description: Azure AD block result
      properties:
        Action:
          type: boolean
          description: Whether the block action was successful
        IOC:
          type: string
          description: The IP address (IOC) attempted to block
        ListName:
          type: string
          description: The Azure AD Named Location list name

    SaseBlockRequest:
      type: object
      properties:
        ioc:
          type: string
          description: Indicator of Compromise
        incident_id:
          type: string
          description: Incident ID
      required: [ioc, incident_id]

    SaseBlockResponse:
      type: object
      description: SASE block result
      properties:
        Added:
          type: boolean
          description: Whether the IOC was added to SASE
        IOC:
          type: string
          description: Indicator of Compromise
        Platform:
          type: string
          description: Platform name (e.g., SASE)
        Action:
          type: string
          description: Action taken (e.g., Block)

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details

security:
  - bearerAuth: []